Finally get the screen to render.
In order to get this working, I end up butchering IGraphicsApi since
I didn't know a better way. This is my first time writing unsafe
code with C#, and my knowledge of C is far from where it should
be, unfortunately :\

There are a few overloads in OpenTK's GL.cs that take in an array
of a generic type, such as BufferData<T>, and I have no idea how
it actually works. Attempting to replicate it results in the
compiler complaining that I can't take a pointer of type T, even
if it restricted to structs, because the compiler doesn't know
for sure that the struct will have only unmanaged fields. Thus,
I just created overloads for all the relevant methods and used
those instead, which is a crappy solution, but it works for now.
I'm not sure how the InteropPatcher gets around that restriction.
A side effect is that the DirectX build is broken.

Also, I converted any internal members of OpenTK's KeyboardDevice
to public so that SDL2Window can use it. Ideally, we wouldn't be
using anything from OpenTK at all, so this is temporary.

Input doesn't work right now, however, so enjoy staring at the
pause menu forever.