Implement some window-related functions as well as keyboard events.
The client now knows about the height and width of the window, and
the window itself can be resized. Also, the current focus state of
the window is known as well. The cursor's visibility can also be
toggled.

As for the keyboard, only the A key actually works right now, but
the framework is there. Ideally, we should not actually use
OpenTK's Key enum anymore; we should prefer SDL2's equivalent instead.