Handle mouse motion and implement mouse-related properties.
The Camera class has its own mouse grabbing mechanism that depends
on PointToScreen and DesktopCursorPos being implemented. When the
latter property is assigned to, the Camera expects the changes to
apply immediately, but the mouse cursor position doesn't change
until the next call to SDL_PumpEvents, which SDL_PollEvent calls.
Thus, in the setter for DesktopCursorPos, we have to call
SDL_PumpEvents right after we warp the mouse cursor, or else the
player's camera is basically unmovable because when the Camera's
previous property is assigned to, it ends up getting the non-centered
position of the cursor.