Change the line ending of OpenGLApi.cs to CRLF.
This is a workaround for the whole file being a merge conflict
every time a change occurs in this file in upstream. I suspect that
the file is being committed with CRLF endings, but my global
core.autocrlf setting is 'input', which converts any changed files
with CRLF endings to have LF endings instead.

Thus, I set my core.autocrlf setting to 'false' for this repository
only as a temporary measure. Having a .gitattributes file in the root
of the repository is ideal, though.