Fix the block being held not rendering initially.
This only occurs in the SDL2 branch because the master branch, which
uses OpenTK, works as intended.

When using OpenTK, UpdateProjection() is called twice while the client
is starting up because the GameWindow calls both OnLoad() and
OnResize() before entering the main loop, and those two methods call
UpdateProjection(). Meanwhile, the SDL2 implementation only called
OnLoad() and not OnResize() since a resize event isn't raised upon
creation of a window in SDL2.

This is important because the first call to UpdateProjection() that
occurs during OnLoad() is performed before the HeldBlockRenderer
object is constructed. When the HeldBlockRenderer object is created,
it adds its delegates to the ProjectionChanged event. Then, when
the resize event is raised by OpenTK and UpdateProjection() is called
again, the HeldBlockRenderer is able to calculate the matrix needed
to render the block being held because the ProjectionChanged event
is raised every time UpdateProjection() is called.

However, since SDL2 doesn't raise the resize event at this point,
the second call doesn't occur, and the ProjectionChanged event isn't
raised. As a result, the Matrix4 property of HeldBlockRenderer will
still consist of zeroes, and the block being held won't show up until
the player does something that causes the projection to be recalculated.

Thus, as a quick fix, mimic OpenTK and call Game.OnResize() before
entering the main loop. What can be investigated is whether the
UpdateProjection() call in OnLoad() can be moved to after the
HeldBlockRenderer is constructed, but I don't know whether that will
cause any other bugs, and I'd rather not change anything unrelated to
SDL2 unless necessary as that can cause merge conflicts later.