Fix a rendering bug in the inventory screen in the DirectX build.
The bug has been present since the changes in
64cb13d0a6522d65f9f615af32e5b009fa815812 were merged.

It can be seen if the player is not hovering over a block when the
inventory screen is being shown. Instead of a blank space where the
block's name would appear, a white rectangle appeared instead.

I overlooked this small change in Direct3D9Api.cs and did not apply
it to the equivalent SDL2 version, which is the cause of that bug.